+++ griptape/drivers/prompt/google_prompt_driver.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from collections.abc import Iterator
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Optional, Any
 from attr import define, field, Factory
 from griptape.utils import PromptStack, import_optional_dependency
 from griptape.artifacts import TextArtifact
@@ -8,8 +8,8 @@ from griptape.drivers import BasePromptDriver
 from griptape.tokenizers import GoogleTokenizer, BaseTokenizer
 
 if TYPE_CHECKING:
-    from google.generativeai.types import ContentDict
     from google.generativeai import GenerativeModel
+    from google.generativeai.types import ContentDict
 
 
 @define
@@ -24,9 +24,9 @@ class GooglePromptDriver(BasePromptDriver):
         top_k: Optional value for top_k.
     """
 
-    api_key: str = field(kw_only=True, metadata={"serializable": True})
+    api_key: Optional[str] = field(default=None, kw_only=True, metadata={"serializable": True})
     model: str = field(kw_only=True, metadata={"serializable": True})
-    model_client: GenerativeModel = field(default=None, kw_only=True)
+    model_client: Any = field(default=Factory(lambda self: self._default_model_client(), takes_self=True), kw_only=True)
     tokenizer: BaseTokenizer = field(
         default=Factory(lambda self: GoogleTokenizer(api_key=self.api_key, model=self.model), takes_self=True),
         kw_only=True,
@@ -34,22 +34,16 @@ class GooglePromptDriver(BasePromptDriver):
     top_p: Optional[float] = field(default=None, kw_only=True, metadata={"serializable": True})
     top_k: Optional[int] = field(default=None, kw_only=True, metadata={"serializable": True})
 
-    def __attrs_post_init__(self):
-        genai = import_optional_dependency("google.generativeai")
-        genai.configure(api_key=self.api_key)
-
-        self.model_client = genai.GenerativeModel(self.model)
-
     def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
         GenerationConfig = import_optional_dependency("google.generativeai.types").GenerationConfig
 
-        input = self._prompt_stack_to_model_input(prompt_stack)
-        chat = self.model_client.start_chat(history=input["history"])
+        history, message = self._prompt_stack_to_model_input(prompt_stack)
+        chat = self.model_client.start_chat(history=history)
         response = chat.send_message(
-            input["message"]["parts"][0],
+            message["parts"][0],
             generation_config=GenerationConfig(
                 stop_sequences=self.tokenizer.stop_sequences,
-                max_output_tokens=self.max_output_tokens(input["history"] + [input["message"]]),
+                max_output_tokens=self.max_output_tokens(history + [message]),
                 temperature=self.temperature,
                 top_p=self.top_p,
                 top_k=self.top_k,
@@ -61,14 +55,14 @@ class GooglePromptDriver(BasePromptDriver):
     def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
         GenerationConfig = import_optional_dependency("google.generativeai.types").GenerationConfig
 
-        input = self._prompt_stack_to_model_input(prompt_stack)
-        chat = self.model_client.start_chat(history=input["history"])
+        history, message = self._prompt_stack_to_model_input(prompt_stack)
+        chat = self.model_client.start_chat(history=history)
         response = chat.send_message(
-            input["message"]["parts"][0],
+            message["parts"][0],
             stream=True,
             generation_config=GenerationConfig(
                 stop_sequences=self.tokenizer.stop_sequences,
-                max_output_tokens=self.max_output_tokens(input["history"] + [input["message"]]),
+                max_output_tokens=self.max_output_tokens(history + [message]),
                 temperature=self.temperature,
                 top_p=self.top_p,
                 top_k=self.top_k,
@@ -78,26 +72,29 @@ class GooglePromptDriver(BasePromptDriver):
         for chunk in response:
             yield TextArtifact(value=chunk.text)
 
-    def default_prompt_stack_to_string_converter(self, prompt_stack: PromptStack) -> str:
-        return "".join([i.content for i in prompt_stack.inputs])
+    def _default_model_client(self) -> GenerativeModel:
+        genai = import_optional_dependency("google.generativeai")
+        genai.configure(api_key=self.api_key)
 
-    def _prompt_stack_to_model_input(self, prompt_stack: PromptStack) -> dict[str, Any]:
+        return genai.GenerativeModel(self.model)
+
+    def _prompt_stack_to_model_input(self, prompt_stack: PromptStack) -> tuple[list[ContentDict], ContentDict]:
         history_inputs = prompt_stack.inputs[:-1]
         message_input = self.__to_content_dict(prompt_stack.inputs[-1])
 
-        content = []
+        history = []
         for prompt_input in history_inputs:
             if prompt_input.is_system():
-                content.append(self.__to_content_dict(prompt_input))
-                content.append(
+                history.append(self.__to_content_dict(prompt_input))
+                history.append(
                     self.__to_content_dict(PromptStack.Input(role=PromptStack.ASSISTANT_ROLE, content="Understood."))
                 )
             elif prompt_input.is_assistant():
-                content.append(self.__to_content_dict(prompt_input))
+                history.append(self.__to_content_dict(prompt_input))
             else:
-                content.append(self.__to_content_dict(prompt_input))
+                history.append(self.__to_content_dict(prompt_input))
 
-        return {"message": message_input, "history": content}
+        return history, message_input
 
     def __to_content_dict(self, prompt_input: PromptStack.Input) -> ContentDict:
         ContentDict = import_optional_dependency("google.generativeai.types").ContentDict